refactor: 메모 응답 데이터 파싱 로직을 Mapper로 분리
- MemoDataSourceImpl에 포함된 Firebase 응답 파싱 코드를 별도의 MemoResponseMapper 객체로 이동하여 관심사 분리 및 가독성 개선
